Limit panic on large or corrupted string length
When we receive a string with a huge length (because it is corrupted or because it is very large), the code will panic.

This patch assume that is the string length is larger than MaxInt32, something is wrong and will return an error.

It should protect the library against most panic raised in that situation.