use Go modules
This PR changes gogen-avro to use Go modules for dependency management.
Because gogen-avro is already tagged with a version >v1, this requires
a new import path, which is a backwardly incompatible change, so to avoid
breaking importers that are tracking master, we move to a `v7` version.

This PR will need to be retargeted as a new `v7` branch to avoid the above-mentioned
breakage.