Implement optional unions as pointers to structs
Unions with optional fields are declared as pointers instead of value structs. Optionality is specified by indicating the position of the null field during creation. If the union is not to be declared optional, the optional field position indicator is a negative number. IsOptional will return true if the position indicator is non-negative. Optional unions are generated as '&UnionTypeName{}' instead of 'UnionTypeName{}' by the constructor method.

Additionally the union has to distinguish between all its avro types (e.g. for generating the composite field name or returning its children) and the actually generated types returned by ItemTypes() to the code generator.