Implement 'reject' switch case instruction, so optional fields can be cleared
As it is designed in the current VM, whenever a union is about to be deserialized, the VM 'enters' into the current target's field-to-be-read. This triggers the creation of the field in advance, which is OK for non-optional fields, but not so OK for optional fields, which must be deserialized as nil pointers instead. Since the field has to exist prior to be deserialized, the VM needs to know how to rollback the creation, that is, how to 'reject' the field at all. The easiest way to tell the VM to reject a field is to use the current EXIT instruction, along with a special operand. The pair (EXIT, NULL) is the one I chose to indicate the rejection of the current field, and (EXIT, NOOP) is the regular, accepting case.

A new switch-case instruction keeps track of the rejection cases, which in turn generates the pair (EXIT, NULL) instruction. Since all switch cases "close" the previous case by generating an initial JUMP to the end of the block (so the last one simply goes on directly to this end), and the rejection switch-case is actually a full rollback, the switch-case generated code must be changed as follows:

  - First swith-case does not generate an initial JUMP, because this instruction simply will never be executed.
  - Any rejection case 'ends' the reading, so they don't need to be closed (completed) with an initial JUMP by the following case.
  - Follow-up cases must generate the initial JUMP to close (complete) the previous case, as long as the previous case needs closing.

The size of each case is now varying from 0 (initial case with readerID -1) to 3 (initial JUMP and case code). Additionally, rejection cases actually EXIT their block, so simply JUMPing to the end of the block will not work, since the end of the block is an accepting EXIT instruction. Rejection cases must create a JUMP one instruction past the end of the block, to overcome the accepting EXIT.

The changes in code sizes and JUMP positions require the JUMP offset to be relative for switch-case instructions. This relative offset needs to be added to the final, computed offset in order to compute the real absolute offset. So jumping to the end of the block generates a (JUMP, 0) instruction and jumping one past the end of the block, as in rejecting cases, generates a (JUMP, 1) instruction.