Implement (EXIT, NULL) in evaluator
'evalInner' now returns a boolean flag 'accepted' to indicate the caller if the read data is accepted or not for optional fields. If accepted is false, then the data must be discarded. This flag is honored only in the ENTER and APPENDMAP instructions, since currently only unions and map values can be nillable. Data is discarded in records by calling its Clear method. Maps are cleared by calling its ClearMap method.

Minor changes:
  - removed unnecessary 'break' from switches.
  - added '// Do nothing' comment in Null operation to make it clear it is empty code.
  - removed global err variable.