Remove ClearMap()
Due to current implementation of map deserialization, the key that would be removed is always the last added one. So there is no need of a ClearMap(string) method, as long as the key/value to be removed is known in runtime. The existing Clear(i int) can be reused to reduce the number of functions.